fix: remove sized bound on byte(reader|writer)
This commit expands the number of types that can implement `ByteReader`
and `ByteWriter` by removing the `Sized` bound on those types, including
on the methods of `Serializable` and `Deserializable` that take one or
the other as arguments.

This makes it possible to implement these traits on a wider variety of
types, and there is no existing need for the Sized bound in winterfell.

This is related to the implementation of `ByteReader` and `ByteWriter`
on standard library types, as some implementations might not be `Sized`.